discord-bots-python-backup-20251225/bots/spelling-bee-bot/requirements.txt
discord.py>=2.0.0
openai>=0.27.0
python-dotenv>=0.19.0
